Added ScoreboardUtils + changed few smaller things
