+++ src/main/java/me/partlysanestudios/partlysaneskies/PartlySaneSkies.java
@@ -21,7 +21,6 @@ package me.partlysanestudios.partlysaneskies;
 
 import cc.polyfrost.oneconfig.config.core.OneColor;
 import gg.essential.elementa.ElementaVersion;
-import ibxm.Player;
 import me.partlysanestudios.partlysaneskies.auctionhouse.menu.AuctionHouseGui;
 import me.partlysanestudios.partlysaneskies.chat.ChatAlertsManager;
 import me.partlysanestudios.partlysaneskies.chat.ChatManager;
@@ -48,18 +47,12 @@ import me.partlysanestudios.partlysaneskies.rngdropbanner.DropBannerDisplay;
 import me.partlysanestudios.partlysaneskies.system.*;
 import me.partlysanestudios.partlysaneskies.system.requests.Request;
 import me.partlysanestudios.partlysaneskies.system.requests.RequestsManager;
-import me.partlysanestudios.partlysaneskies.utils.LocationUtils;
-import me.partlysanestudios.partlysaneskies.utils.StringUtils;
 import me.partlysanestudios.partlysaneskies.utils.Utils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.event.ClickEvent;
 import net.minecraft.event.HoverEvent;
 import net.minecraft.inventory.IInventory;
-import net.minecraft.scoreboard.Score;
-import net.minecraft.scoreboard.ScoreObjective;
-import net.minecraft.scoreboard.ScorePlayerTeam;
-import net.minecraft.scoreboard.Scoreboard;
 import net.minecraft.util.ChatComponentText;
 import net.minecraft.util.IChatComponent;
 import net.minecraftforge.client.event.ClientChatReceivedEvent;
@@ -79,10 +72,6 @@ import org.apache.logging.log4j.Logger;
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
 
 
 @Mod(modid = PartlySaneSkies.MODID, version = PartlySaneSkies.VERSION, name = PartlySaneSkies.NAME)
@@ -411,41 +400,11 @@ public class PartlySaneSkies {
         return new IInventory[] { upperInventory, lowerInventory };
     }
 
-    // Returns the name of the scoreboard without color codes
-    public static String getScoreboardName() {
-        String scoreboardName = minecraft.thePlayer.getWorldScoreboard().getObjectiveInDisplaySlot(1).getDisplayName();
-        return StringUtils.removeColorCodes(scoreboardName);
-    }
-
     // Runs when debug key is pressed
     public static void debugMode() {
         PartlySaneSkies.isDebugMode = !PartlySaneSkies.isDebugMode;
         Utils.sendClientMessage("Debug mode: " + PartlySaneSkies.isDebugMode);
         BannerRenderer.INSTANCE.renderNewBanner(new PSSBanner("Test", 5000L, 5f, new OneColor(255, 0, 255, 1).toJavaColor()));
-    }
-
-    // Returns a list of lines on the scoreboard,
-    // where each line is a new entry
-    public static List<String> getScoreboardLines() {
-        try {
-            Scoreboard scoreboard = minecraft.theWorld.getScoreboard();
-            ScoreObjective objective = scoreboard.getObjectiveInDisplaySlot(1);
-            Collection<Score> scoreCollection = scoreboard.getSortedScores(objective);
-
-            List<String> scoreLines = new ArrayList<>();
-            for (Score score : scoreCollection) {
-                scoreLines.add(ScorePlayerTeam.formatPlayerName(scoreboard.getPlayersTeam(score.getPlayerName()),
-                        score.getPlayerName()));
-            }
-
-            return scoreLines;
-        } catch (IllegalArgumentException e) {
-            if (e.getMessage().equals("Cannot locate declared field class net.minecraft.client.gui.inventory.GuiChest.field_147015_w")) {
-                System.out.println("Strange error message in PartlySaneSkies.getScoreboardLines()");
-            }
-            e.printStackTrace();
-            return Collections.emptyList();
-        }
     }
 
 
@@ -454,79 +413,6 @@ public class PartlySaneSkies {
         return System.currentTimeMillis();
     }
 
-
-    // Gets the number of coins in your purse from the scoreboard
-    public static long getCoins() {
-        if (!LocationUtils.isSkyblock()) {
-            return 0L;
-        }
-
-        List<String> scoreboard = getScoreboardLines();
-
-        String money = null;
-
-        for (String line : scoreboard) {
-            if (StringUtils.stripLeading(line).contains("Piggy:") || StringUtils.stripLeading(line).contains("Purse:")) {
-                money = StringUtils.stripLeading(StringUtils.removeColorCodes(line)).replace("Piggy: ", "");
-                money = StringUtils.stripLeading(StringUtils.removeColorCodes(money)).replace("Purse: ", "");
-                money = StringUtils.stripLeading(money);
-                money = money.replace(",", "");
-                money = money.replaceAll("\\P{Print}", "");
-                break;
-            }
-        }
-
-        if (money == null) {
-            return 0L;
-        }
-        try {
-            return Long.parseLong(money);
-        } catch (NumberFormatException event) {
-            return 0;
-        }
-    }
-
-    // Gets the number of bits from the scoreboard
-    public static long getBits() {
-        if (!LocationUtils.isSkyblock()) {
-            return 0L;
-        }
-
-        List<String> scoreboard = getScoreboardLines();
-
-        String bits = null;
-
-        for (String line : scoreboard) {
-            if (StringUtils.stripLeading(line).contains("Bits:")) {
-                bits = StringUtils.stripLeading(StringUtils.removeColorCodes(line)).replace("Bits: ", "");
-                bits = StringUtils.stripLeading(bits);
-                bits = bits.replace(",", "");
-                bits = bits.replaceAll("\\P{Print}", "");
-                break;
-            }
-        }
-
-        if (bits == null) {
-            return 0L;
-        }
-
-        String[] charsToRemove = {"(", ")", ".", "-", "+"};
-
-        for (String removalChar : charsToRemove) {
-            if (bits.contains(removalChar)) {
-                int indexOfEndOfCount = bits.indexOf(removalChar);
-                bits = bits.substring(0, indexOfEndOfCount);
-            }
-        }
-
-        bits = StringUtils.stripLeading(bits);
-        bits = StringUtils.stripTrailing(bits);
-        try {
-            return Long.parseLong(bits);
-        } catch (NumberFormatException event) {
-            return 0;
-        }
-    }
 
     // Sends a ping to the count API to track the number of users per day
     public void trackLoad() {
